requirements.txt
alabaster==0.7.12
autopep8==1.4.4
Babel==2.6.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
defusedxml==0.6.0
Django==2.2
django-appconf==1.0.3
django-bootstrap-breadcrumbs==0.9.1
django-bower==5.2.0
django-braces==1.13.0
django-compressor==2.2
django-cors-headers==3.0.1
django-debug-toolbar==1.11
django-filter==2.1.0
django-filters==0.2.1
django-formtools==2.1
django-guardian==1.5.1
django-mobile==0.7.0
django-oauth-toolkit==1.2.0
django-recaptcha==2.0.4
django-recaptcha2==1.4.1
django-rest-framework-social-oauth2==1.1.0
django-social-share==1.3.2
django-sortedm2m==1.5.0
django-tastypie==0.14.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
docutils==0.14
easy-thumbnails==2.6
entrypoints==0.3
flake8==3.7.7
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-storage==1.16.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gunicorn==19.9.0
hound==0.1.8
icalendar==4.0.3
idna==2.8
imagesize==1.1.0
invoke==1.2.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
numpy==1.16.4
oauthlib==3.0.1
packaging==19.0
pandas==0.24.2
pep8==1.7.1
Pillow==6.0.0
protobuf==3.8.0
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.1
PyJWT==1.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2019.1
rcssmin==1.0.6
reportlab==3.5.21
requests==2.22.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
rsa==4.0
six==1.12.0
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
urllib3==1.25.3
webencodings==0.5.1
-e git+https://github.com/andela/wg-baqi.git@e85134cce08c229af979f8437acfde8b51ca925c#egg=wger
whitenoise==4.1.2
